refactor(library): Move StreetViewJavaHelper to library module
This commit moves the `StreetViewHelper` from the `demo` module to the `library` module, making it an officially supported utility for Java consumers.

The key changes include:
- Renaming `StreetViewHelper` to `StreetViewJavaHelper` for better clarity and moving it to the `com.google.maps.android` package within the library.
- Migrating the corresponding unit tests from the `demo` to the `library` module and updating them to reflect the class rename.
- Adjusting `build.gradle.kts` files to move test dependencies (`mockk`, `coroutines-test`, `robolectric`) from the `demo` to the `library` module.
- Renaming `StreetViewDemoActivityJava` to `StreetViewDemoJavaActivity` and updating it to use the new helper from the library.
- Improving the warning message for invalid or missing API keys in the demo app.
