+++ demo/src/main/res/values/strings.xml
@@ -33,5 +33,5 @@
     <string name="button_radius">Radius</string>
     <string name="button_gradient">Gradient</string>
     <string name="button_opacity">Opacity</string>
-    <string name="bad_maps_api_key">Invalid or missing Google Maps API key</string>
+    <string name="bad_maps_api_key">The Google Maps API key looks invalid or is missing. If you are sure your key is correct, you can remove this check from the demo.</string>
 </resources>
