+++ library/build.gradle.kts
@@ -71,6 +71,10 @@ dependencies {
     testImplementation(libs.kotlin.test)
     testImplementation(libs.truth)
     implementation(libs.kotlin.stdlib.jdk8)
+
+    testImplementation(libs.mockk)
+    testImplementation(libs.kotlinx.coroutines.test)
+    testImplementation(libs.robolectric)
 }
 
 tasks.register("instrumentTest") {
