+++ library/src/main/java/com/google/maps/android/StreetViewJavaHelper.kt
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,12 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package com.google.maps.android.utils.demo
+package com.google.maps.android
 
 import com.google.android.gms.maps.model.LatLng
-import com.google.maps.android.StreetViewUtils
-import com.google.maps.android.Status
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -26,7 +23,7 @@ import kotlinx.coroutines.launch
 /**
  * A helper object to call the suspend function `fetchStreetViewData` from Java.
  */
-object StreetViewHelper {
+object StreetViewJavaHelper {
     /**
      * A callback interface to receive the result of the Street View data fetch.
      */
